Moved timestamp-related code to program/timestamp.rs
