+++ rust/template/differential_datalog/program/mod.rs
@@ -12,24 +12,24 @@
 // TODO: namespace cleanup
 // TODO: single input relation
 
+mod timestamp;
+
+pub use timestamp::{TSNested, TupleTS, TS, TS16};
+
+use crate::{ddval::*, profile::*, record::Mutator, variable::*};
+use fnv::{FnvBuildHasher, FnvHashMap, FnvHashSet};
 use std::{
     collections::{hash_map, BTreeMap, BTreeSet, HashMap},
     fmt::{self, Debug, Formatter},
-    ops::{Add, Deref, Mul},
+    ops::{Deref, Mul},
     sync::{
-        atomic::{AtomicBool, AtomicU32, Ordering},
+        atomic::{AtomicBool, Ordering},
         mpsc, Arc, Barrier, Mutex,
     },
     thread,
     time::Duration,
 };
-
-use abomonation::Abomonation;
-
-// use deterministic hash-map and hash-set, as differential dataflow expects deterministic order of
-// creating relations
-use fnv::{FnvBuildHasher, FnvHashMap, FnvHashSet};
-use num::One;
+use timestamp::{TSAtomic, ToTupleTS};
 
 use differential_dataflow::difference::Diff;
 use differential_dataflow::difference::Monoid;
@@ -55,12 +55,10 @@ use timely::dataflow::operators::*;
 use timely::dataflow::scopes::*;
 use timely::dataflow::ProbeHandle;
 use timely::logging::TimelyEvent;
-use timely::order::{PartialOrder, Product, TotalOrder};
+use timely::order::{Product, TotalOrder};
 use timely::progress::timestamp::Refines;
-use timely::progress::{PathSummary, Timestamp};
+use timely::progress::Timestamp;
 use timely::worker::Worker;
-
-use crate::{ddval::*, profile::*, record::Mutator, variable::*};
 
 type ValTrace<S> = DefaultValTrace<DDValue, DDValue, <S as ScopeParent>::Timestamp, Weight, u32>;
 type KeyTrace<S> = DefaultKeyTrace<DDValue, <S as ScopeParent>::Timestamp, Weight, u32>;
@@ -70,127 +68,6 @@ type TKeyAgent<S> = TraceAgent<KeyTrace<S>>;
 
 type TValEnter<'a, P, T> = TraceEnter<TValAgent<P>, T>;
 type TKeyEnter<'a, P, T> = TraceEnter<TKeyAgent<P>, T>;
-
-/// 16-bit timestamp.
-// TODO: get rid of this and use `u16` directly when/if differential implements
-// `Lattice`, `Timestamp`, `PathSummary` traits for `u16`.
-#[derive(Copy, PartialOrd, PartialEq, Eq, Debug, Default, Clone, Hash, Ord)]
-pub struct TS16 {
-    pub x: u16,
-}
-
-impl Abomonation for TS16 {}
-
-impl Mul for TS16 {
-    type Output = TS16;
-    fn mul(self, rhs: TS16) -> Self::Output {
-        TS16 { x: self.x * rhs.x }
-    }
-}
-
-impl Add for TS16 {
-    type Output = TS16;
-
-    fn add(self, rhs: TS16) -> Self::Output {
-        TS16 { x: self.x + rhs.x }
-    }
-}
-
-impl One for TS16 {
-    fn one() -> Self {
-        TS16 { x: 1 }
-    }
-}
-
-impl TS16 {
-    pub const fn max_value() -> TS16 {
-        TS16 { x: 0xffff }
-    }
-}
-
-impl PartialOrder for TS16 {
-    fn less_equal(&self, other: &Self) -> bool {
-        self.x.less_equal(&other.x)
-    }
-
-    fn less_than(&self, other: &Self) -> bool {
-        self.x.less_than(&other.x)
-    }
-}
-
-impl Lattice for TS16 {
-    fn minimum() -> Self {
-        TS16 {
-            x: u16::min_value(),
-        }
-    }
-
-    fn join(&self, other: &Self) -> Self {
-        TS16 {
-            x: std::cmp::max(self.x, other.x),
-        }
-    }
-
-    fn meet(&self, other: &Self) -> Self {
-        TS16 {
-            x: std::cmp::min(self.x, other.x),
-        }
-    }
-}
-
-impl Timestamp for TS16 {
-    type Summary = TS16;
-}
-
-impl PathSummary<TS16> for TS16 {
-    fn results_in(&self, src: &TS16) -> Option<TS16> {
-        self.x.checked_add(src.x).map(|y| TS16 { x: y })
-    }
-
-    fn followed_by(&self, other: &TS16) -> Option<TS16> {
-        self.x.checked_add(other.x).map(|y| TS16 { x: y })
-    }
-}
-
-impl From<TS16> for u64 {
-    fn from(ts: TS16) -> Self {
-        ts.x as u64
-    }
-}
-
-/// Outer timestamp
-pub type TS = u32;
-type TSAtomic = AtomicU32;
-
-/// Timestamp for the nested scope
-/// Use 16-bit timestamps for inner scopes to save memory
-#[cfg(feature = "nested_ts_32")]
-pub type TSNested = u32;
-
-/// Timestamp for the nested scope
-/// Use 16-bit timestamps for inner scopes to save memory
-#[cfg(not(feature = "nested_ts_32"))]
-pub type TSNested = TS16;
-
-/// `Inspect` operator expects the timestampt to be a tuple.
-pub type TupleTS = (TS, TSNested);
-
-trait ToTupleTS {
-    fn to_tuple_ts(&self) -> TupleTS;
-}
-
-/// 0-extend top-level timestamp to a tuple.
-impl ToTupleTS for TS {
-    fn to_tuple_ts(&self) -> TupleTS {
-        (*self, TSNested::default())
-    }
-}
-
-impl ToTupleTS for Product<TS, TSNested> {
-    fn to_tuple_ts(&self) -> TupleTS {
-        (self.outer, self.inner)
-    }
-}
 
 /// Diff associated with records in differential dataflow
 pub type Weight = i32;
